Rebase problems

+++ cmd/dag_analyzer/main.go
@@ -13,7 +13,7 @@ import (
 type dagFactory struct{}
 
 func (dagFactory) CreateDag(dc gomel.DagConfig) gomel.Dag {
-	return dag.New(len(dc.Keys))
+	return dag.New(uint16(len(dc.Keys)))
 }
 
 type cliOptions struct {
